+++ src/library_dylink.js
@@ -390,11 +390,6 @@ var LibraryDylink = {
       };
       var proxy = new Proxy(env, proxyHandler);
       var info = {
-        'global': {
-          'NaN': NaN,
-          'Infinity': Infinity,
-        },
-        'global.Math': Math,
         'GOT.mem': new Proxy(asmLibraryArg, GOTHandler),
         'GOT.func': new Proxy(asmLibraryArg, GOTHandler),
         'env': proxy,
@@ -621,7 +616,7 @@ var LibraryDylink = {
 
   $preloadDylibs__deps: ['$loadDynamicLibrary', '$reportUndefinedSymbols'],
   $preloadDylibs: function() {
-#ifdef DYLINK_DEBUG
+#if DYLINK_DEBUG
     err('preloadDylibs');
 #endif
     var libs = {{{ JSON.stringify(RUNTIME_LINKED_LIBS) }}};
@@ -629,7 +624,7 @@ var LibraryDylink = {
       libs = libs.concat(Module['dynamicLibraries'])
     }
     if (!libs.length) {
-#ifdef DYLINK_DEBUG
+#if DYLINK_DEBUG
       err('preloadDylibs: no libraries to preload');
 #endif
       reportUndefinedSymbols();